% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectRdkit.R
\name{callPythonScript}
\alias{callPythonScript}
\title{python system call}
\usage{
callPythonScript(
  inputVec,
  pyfile,
  pythonPath = getOption("dartpaths_python"),
  header = TRUE,
  outputClass = "data.table",
  useFiles = FALSE
)
}
\arguments{
\item{inputVec}{single string or a vector with strings to be sent to python script via stdin}

\item{pyfile}{python file to be executed via system call}

\item{pythonPath}{full path of python executable}

\item{header}{If TRUE, the first array in the returned json is used as a header for the returned table}

\item{outputClass}{Class name of returned object (default: data.table)}

\item{useFiles}{If TRUE, files are used to exchange files between python and R. If FALSE, stdin/stdout is used.}
}
\value{
data.table object
}
\description{
use system call to apply Rdkit normalization to SMILES representation of chemical structures
}
\author{
Marvin Steijaert
}
