% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substanceUtils.R
\name{CAStoInt}
\alias{CAStoInt}
\title{Function to convert character representation of CAS numbers to integer representation}
\usage{
CAStoInt(vec)
}
\arguments{
\item{vec}{character vector with CAS numbers}
}
\value{
integer vector with CAS numbers
}
\description{
Function to convert character representation of CAS numbers to integer representation
}
\author{
Marvin Steijaert
}
