% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substanceUtils.R
\name{clusterByMatchingEntries}
\alias{clusterByMatchingEntries}
\title{Multi-column clustering}
\usage{
clusterByMatchingEntries(
  inputTable,
  clusterColumns,
  uniqueColumns,
  verbose = FALSE
)
}
\arguments{
\item{inputTable}{data.table object}

\item{clusterColumns}{Columns for which at least one item should match}

\item{uniqueColumns}{Columns for which all items should match}

\item{verbose}{Toggle verbosity}
}
\description{
Function that allows clustering of rows in a table that have matching values for one the columns
}
\details{
An additional column 'cluster' is added to the input table (modified in-place)
NA values in clusterColumns are ignored
}
\author{
Marvin Steijaert
}
