% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createAsymmetricTanimotoMatrix}
\alias{createAsymmetricTanimotoMatrix}
\title{createAsymmetricTanimotoMatrix}
\usage{
createAsymmetricTanimotoMatrix(fingerprintsA, fingerprintsB)
}
\arguments{
\item{fingerprintsA}{Matrix with 0/1 values (rows: compounds, columns: fingerprint features)}

\item{fingerprintsB}{Matrix with 0/1 values (rows: compounds, columns: fingerprint features)}
}
\value{
matrix with Tanimoto similarities
}
\description{
Function to calculate all pairwise Tanimoto similarities between two sets of fingerprints
}
\author{
Marvin Steijaert
}
