% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appCompound.R
\name{createBarImg}
\alias{createBarImg}
\title{Create a colored bar indicating percentages}
\usage{
createBarImg(nLeft, nRight, colorOrder = "magentagreen")
}
\arguments{
\item{nLeft}{Number of left color counts}

\item{nRight}{Number of right color counts}

\item{colorOrder}{Order of colors. Available options: "magentagreen"(default), "yellowblue", "redgreen", "greenred"}
}
\value{
string with html img element containing base64 encoded image
}
\description{
Create a colored bar indicating percentages
}
\examples{
\dontrun{createBarImg(50,50)}
}
\author{
Marvin Steijaert
}
