% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createContactButton.R
\name{createContactButton}
\alias{createContactButton}
\title{Create contact buttons}
\usage{
createContactButton(text)
}
\arguments{
\item{text}{text (character vector of length one)}
}
\description{
Create contact buttons
}
