% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appGene.R
\name{createDisabledLink}
\alias{createDisabledLink}
\title{Helper function: Create disabled button for 0 label items}
\usage{
createDisabledLink(label)
}
\arguments{
\item{label}{character, label shown in table}
}
\value{
dedicated div
}
\description{
Helper function: Create disabled button for 0 label items
}
