% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appGene.R
\name{createLink}
\alias{createLink}
\title{Helper function: Create link to webpage in datatable}
\usage{
createLink(webLink, label)
}
\arguments{
\item{webLink}{character, web address}

\item{label}{character, label shown in table}
}
\value{
weblink refering to page \code{webLink} with label \code{description}
}
\description{
Helper function: Create link to webpage in datatable
}
\author{
mvarewyck
}
