% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appGene.R
\name{createModalLink}
\alias{createModalLink}
\title{Helper function: Create actionLink to open modalDialog in datatable}
\usage{
createModalLink(
  species,
  row,
  label,
  colors = list(`1:1` = "#bee126", `1:N` = "#ffde17", `N:N` = "#488df4;")
)
}
\arguments{
\item{species}{character, species for which to create link}

\item{row}{integer, row in datatable}

\item{label}{character, label shown in table}

\item{colors}{list, colors for the gene classes 
\code{c("one2one", "one2many", "many2many")} (optional)}
}
\value{
actionLink with identifier including the \code{species} and table \code{row}
}
\description{
Helper function: Create actionLink to open modalDialog in datatable
}
\author{
mvarewyck
}
