% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createTanimotoMatrix}
\alias{createTanimotoMatrix}
\title{createTanimotoMatrix}
\usage{
createTanimotoMatrix(fingerprints, oldMatrix = matrix())
}
\arguments{
\item{fingerprints}{Matrix with 0/1 values (rows: compounds, columns: fingerprint features)}

\item{oldMatrix}{Optional numeric matrix with precalculated similarities. The corresponding part of the
result matrix will be copied, rather than recalculated Logical, indicates wheter oldMatrix should be used}
}
\value{
square matrix with Tanimoto similarities
}
\description{
Function to calculate all pairwise Tanimoto similarities for one set of fingerprints
}
\examples{
a=matrix(sample(0:1,100,TRUE),10,10);b=matrix(1:4,2,2);createTanimotoMatrix(a,b)
}
\author{
Marvin Steijaert
}
