% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDartDB.R
\name{DartDB}
\alias{DartDB}
\title{Class for storing DARTpaths data}
\description{
Class for storing DARTpaths data

Class for storing DARTpaths data
}
\details{
For reasons of code visibility, the different components are implemented in a series of parent classes
}
\section{Super classes}{
\code{\link[dartTools:Data]{dartTools::Data}} -> \code{\link[dartTools:Substances]{dartTools::Substances}} -> \code{\link[dartTools:PhenoPathway]{dartTools::PhenoPathway}} -> \code{DartDB}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{classname}}{Name of the class}

\item{\code{tableColumns}}{Named list with names of all the columns for each table}

\item{\code{columnConversionFunctions}}{Named vector with functions (e.g. as.character) that applied after loading a columns with that name in any of the tables}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{DartDB$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="addData">}\href{../../dartTools/html/Data.html#method-addData}{\code{dartTools::Data$addData()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="addTable">}\href{../../dartTools/html/Data.html#method-addTable}{\code{dartTools::Data$addTable()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="checkTableConventions">}\href{../../dartTools/html/Data.html#method-checkTableConventions}{\code{dartTools::Data$checkTableConventions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="dump">}\href{../../dartTools/html/Data.html#method-dump}{\code{dartTools::Data$dump()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="getData">}\href{../../dartTools/html/Data.html#method-getData}{\code{dartTools::Data$getData()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="initialize">}\href{../../dartTools/html/Data.html#method-initialize}{\code{dartTools::Data$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="loadDump">}\href{../../dartTools/html/Data.html#method-loadDump}{\code{dartTools::Data$loadDump()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="ncols">}\href{../../dartTools/html/Data.html#method-ncols}{\code{dartTools::Data$ncols()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="nrows">}\href{../../dartTools/html/Data.html#method-nrows}{\code{dartTools::Data$nrows()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="print">}\href{../../dartTools/html/Data.html#method-print}{\code{dartTools::Data$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="printAllTables">}\href{../../dartTools/html/Data.html#method-printAllTables}{\code{dartTools::Data$printAllTables()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="writeTxtTable">}\href{../../dartTools/html/Data.html#method-writeTxtTable}{\code{dartTools::Data$writeTxtTable()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id=".mergeSubstancesOnce">}\href{../../dartTools/html/Substances.html#method-.mergeSubstancesOnce}{\code{dartTools::Substances$.mergeSubstancesOnce()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id=".updateSubstanceids">}\href{../../dartTools/html/Substances.html#method-.updateSubstanceids}{\code{dartTools::Substances$.updateSubstanceids()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="addConstituents">}\href{../../dartTools/html/Substances.html#method-addConstituents}{\code{dartTools::Substances$addConstituents()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="addIdTypes">}\href{../../dartTools/html/Substances.html#method-addIdTypes}{\code{dartTools::Substances$addIdTypes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="addSubstanceActivity">}\href{../../dartTools/html/Substances.html#method-addSubstanceActivity}{\code{dartTools::Substances$addSubstanceActivity()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="addSubstanceCategory">}\href{../../dartTools/html/Substances.html#method-addSubstanceCategory}{\code{dartTools::Substances$addSubstanceCategory()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="addSubstances">}\href{../../dartTools/html/Substances.html#method-addSubstances}{\code{dartTools::Substances$addSubstances()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getDescriptors">}\href{../../dartTools/html/Substances.html#method-getDescriptors}{\code{dartTools::Substances$getDescriptors()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getMonoSubstanceSmiles">}\href{../../dartTools/html/Substances.html#method-getMonoSubstanceSmiles}{\code{dartTools::Substances$getMonoSubstanceSmiles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getSameCategorySubstances">}\href{../../dartTools/html/Substances.html#method-getSameCategorySubstances}{\code{dartTools::Substances$getSameCategorySubstances()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getSimilarConstituents">}\href{../../dartTools/html/Substances.html#method-getSimilarConstituents}{\code{dartTools::Substances$getSimilarConstituents()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getSimilarSubstances">}\href{../../dartTools/html/Substances.html#method-getSimilarSubstances}{\code{dartTools::Substances$getSimilarSubstances()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getSubstanceActivity">}\href{../../dartTools/html/Substances.html#method-getSubstanceActivity}{\code{dartTools::Substances$getSubstanceActivity()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getSubstanceExtIds">}\href{../../dartTools/html/Substances.html#method-getSubstanceExtIds}{\code{dartTools::Substances$getSubstanceExtIds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getSubstanceIds">}\href{../../dartTools/html/Substances.html#method-getSubstanceIds}{\code{dartTools::Substances$getSubstanceIds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="inspect">}\href{../../dartTools/html/Substances.html#method-inspect}{\code{dartTools::Substances$inspect()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="mapSmilesToSubstance">}\href{../../dartTools/html/Substances.html#method-mapSmilesToSubstance}{\code{dartTools::Substances$mapSmilesToSubstance()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="maxSubstanceid">}\href{../../dartTools/html/Substances.html#method-maxSubstanceid}{\code{dartTools::Substances$maxSubstanceid()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="mergeSubstances">}\href{../../dartTools/html/Substances.html#method-mergeSubstances}{\code{dartTools::Substances$mergeSubstances()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="tabulateSubstanceCategory">}\href{../../dartTools/html/Substances.html#method-tabulateSubstanceCategory}{\code{dartTools::Substances$tabulateSubstanceCategory()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="tabulateSubstanceInfo">}\href{../../dartTools/html/Substances.html#method-tabulateSubstanceInfo}{\code{dartTools::Substances$tabulateSubstanceInfo()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="updateDescriptors">}\href{../../dartTools/html/Substances.html#method-updateDescriptors}{\code{dartTools::Substances$updateDescriptors()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="PhenoPathway" data-id="getHumanPathwayGenes">}\href{../../dartTools/html/PhenoPathway.html#method-getHumanPathwayGenes}{\code{dartTools::PhenoPathway$getHumanPathwayGenes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="PhenoPathway" data-id="getHumanPathwayLevels">}\href{../../dartTools/html/PhenoPathway.html#method-getHumanPathwayLevels}{\code{dartTools::PhenoPathway$getHumanPathwayLevels()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="PhenoPathway" data-id="getHumanPathwayParents">}\href{../../dartTools/html/PhenoPathway.html#method-getHumanPathwayParents}{\code{dartTools::PhenoPathway$getHumanPathwayParents()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="PhenoPathway" data-id="getHumanPathways">}\href{../../dartTools/html/PhenoPathway.html#method-getHumanPathways}{\code{dartTools::PhenoPathway$getHumanPathways()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="PhenoPathway" data-id="getHumanPathwaysForLevel">}\href{../../dartTools/html/PhenoPathway.html#method-getHumanPathwaysForLevel}{\code{dartTools::PhenoPathway$getHumanPathwaysForLevel()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="PhenoPathway" data-id="getPathwayOrthology">}\href{../../dartTools/html/PhenoPathway.html#method-getPathwayOrthology}{\code{dartTools::PhenoPathway$getPathwayOrthology()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="PhenoPathway" data-id="mapGenesToHumanGenes">}\href{../../dartTools/html/PhenoPathway.html#method-mapGenesToHumanGenes}{\code{dartTools::PhenoPathway$mapGenesToHumanGenes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="PhenoPathway" data-id="retrieveTaxon">}\href{../../dartTools/html/PhenoPathway.html#method-retrieveTaxon}{\code{dartTools::PhenoPathway$retrieveTaxon()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DartDB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
