% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etlUtils.R
\name{downloadWebOrFtp}
\alias{downloadWebOrFtp}
\title{utility function to download files from web and ftp locations}
\usage{
downloadWebOrFtp(url, directory, filename = NULL, writeLog = TRUE)
}
\arguments{
\item{url}{url}

\item{directory}{output directory}

\item{filename}{name of output file (if NULL, this is automatically generated from url)}

\item{writeLog}{Logical value indicating if a log file is written with original comamnd and timestamp}
}
\description{
utility function to download files from web and ftp locations
}
\author{
Marvin Steijaert
}
