% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substanceUtils.R
\name{ECtoInt}
\alias{ECtoInt}
\title{Function to convert character representation of EC numbers to integer representation}
\usage{
ECtoInt(vec)
}
\arguments{
\item{vec}{character vector with EC numbers}
}
\value{
integer vector with EC numbers
}
\description{
Function to convert character representation of EC numbers to integer representation
}
\author{
Marvin Steijaert
}
