% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classData.R
\name{emptyDT}
\alias{emptyDT}
\title{Function to create a data.table with specified columns and zero rows}
\usage{
emptyDT(columnNames)
}
\arguments{
\item{columnNames}{Vector with column names}
}
\value{
data.table object
}
\description{
Function to create a data.table with specified columns and zero rows
}
\author{
Marvin Steijaert
}
