% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substanceUtils.R
\name{filterInvalid}
\alias{filterInvalid}
\title{Function to filter entries}
\usage{
filterInvalid(vec, type = "smiles", mode = "replace")
}
\arguments{
\item{vec}{Character vector}

\item{type}{Type of entries ("smiles", "names")}

\item{mode}{How to filter. "replace" replaces by NA, "invalid" returns only invalid entries, "valid" return only valid entries}
}
\value{
Vector. Contents depend on mode argument
}
\description{
Function to filter entries
}
\examples{
\dontrun{
  filterInvalid(c("CCCCC","NA"," - "), "smiles")
  filterInvalid(c("oxygen","NA"," - ", "this is a substance description not a name"), "substancenames")
}
}
\author{
Marvin Steijaert
}
