% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectRdkit.R
\name{fingerprintsRdkit}
\alias{fingerprintsRdkit}
\title{calculate Morgan fingerprints}
\usage{
fingerprintsRdkit(smiles, pythonPath = getOption("dartpaths_python"))
}
\arguments{
\item{smiles}{single SMILES string or a vector with multiple SMILES}

\item{pythonPath}{full path of python executable}
}
\value{
data.table object with smiles and fingerprints
}
\description{
use python RDkit package to calculate morgan fingerprints
}
\examples{
\dontrun{
smiles <- c("N[C@H](C)C(=O)O","O1CCCC[C@H]1C")
result <- fingerprintsRdkit(smiles)
}
}
\author{
Marvin Steijaert
}
