% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{getAllJsonElements}
\alias{getAllJsonElements}
\title{get the value for the same key from multiple json string}
\usage{
getAllJsonElements(vec, jsonkey)
}
\arguments{
\item{vec}{vector of json strings}

\item{jsonkey}{key of item}
}
\value{
character string
}
\description{
get the value for the same key from multiple json string
}
\author{
Marvin Steijaert
}
