% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testUtils.R
\name{getExampleData}
\alias{getExampleData}
\title{Utility function to generate example data for unit testing}
\usage{
getExampleData(dataType = "substance")
}
\arguments{
\item{dataType}{Character string with type of vector (currently, only 'substance' is supported)}
}
\value{
data.table object
}
\description{
Utility function to generate example data for unit testing
}
\author{
Marvin Steijaert
}
