% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{getHstoreElement}
\alias{getHstoreElement}
\title{get the value for the specified key from a (vector of) hstore string(s)}
\usage{
getHstoreElement(hstore_str, hstore_key)
}
\arguments{
\item{hstore_str}{vector of hstore strings}

\item{hstore_key}{key}
}
\value{
vector of values
}
\description{
get the value for the specified key from a (vector of) hstore string(s)
}
\author{
Marvin Steijaert
}
