% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{getHstoreKeys}
\alias{getHstoreKeys}
\title{get all keys in a hstore string}
\usage{
getHstoreKeys(hstoreStr)
}
\arguments{
\item{hstoreStr}{hstore string}
}
\value{
vector with keys
}
\description{
get all keys in a hstore string
}
\author{
Marvin Steijaert
}
