% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{getJsonElement}
\alias{getJsonElement}
\title{get value for specified key from json string}
\usage{
getJsonElement(jsonstr, jsonkey)
}
\arguments{
\item{jsonstr}{json string}

\item{jsonkey}{key of item}
}
\value{
character string
}
\description{
get value for specified key from json string
}
\author{
Marvin Steijaert
}
