% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPhenoPathway.R
\name{getRankMax}
\alias{getRankMax}
\title{auxiliary function for pathway ranking}
\usage{
getRankMax(nrowProfile, nrowTargetGenes)
}
\arguments{
\item{nrowProfile}{number of genes associated with observed phenotypes}

\item{nrowTargetGenes}{number of genes in target pathway}
}
\value{
rankMax value
}
\description{
auxiliary function for pathway ranking
}
