% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPathwayRanking.R
\name{harmonicMeanColumnWise}
\alias{harmonicMeanColumnWise}
\title{Return the element-wise harmonic mean of multiple vectors}
\usage{
harmonicMeanColumnWise(..., na.rm = FALSE, adjustMethod = NULL)
}
\arguments{
\item{...}{one or more vectors (all with the same length)}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}

\item{adjustMethod}{Name of applyied multiple testing correction (passed to p.adjust()). If NULL, original values are returned.}
}
\value{
maximum value
}
\description{
Return the element-wise harmonic mean of multiple vectors
}
\details{
For convenience, ... can include NULL values. These are ignored.
}
\examples{
dartTools:::harmonicMeanColumnWise(c(0.1,0.2,0.3), c(0.3, 0.4, 0.5))
}
