% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPathwayRanking.R
\name{harmonicMeanWithNA}
\alias{harmonicMeanWithNA}
\title{Returns NA if only NA values are provided. Otherwise, the harmonic mean of non-NA values is returned}
\usage{
harmonicMeanWithNA(vec)
}
\arguments{
\item{vec}{numeric vector}
}
\value{
maximum value
}
\description{
Returns NA if only NA values are provided. Otherwise, the harmonic mean of non-NA values is returned
}
