% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appGene.R
\name{interactiveGeneTable}
\alias{interactiveGeneTable}
\title{Create interactive table for gene data}
\usage{
interactiveGeneTable(
  geneData,
  colors = list(`1:1` = "#bee126", `1:N` = "#ffde17", `N:N` = "#488df4"),
  opts = NULL,
  columnNames = NULL
)
}
\arguments{
\item{geneData}{data.table object (as returned by \code{DartDB$getPathwayOrthology()})}

\item{colors}{list, colors for the gene classes 
\code{c("one2one", "one2many", "many2many")} (optional)}

\item{opts}{options to be passed to the \code{datatable} downstream}

\item{columnNames}{character vector, column names for the interactive table}
}
\value{
datatable, interactive version of \code{geneData}
}
\description{
Create interactive table for gene data
}
\author{
mvarewyck
}
