% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substanceUtils.R
\name{intToCAS}
\alias{intToCAS}
\title{Function to convert integer representation of CAS numbers to character representation}
\usage{
intToCAS(vec)
}
\arguments{
\item{vec}{integer vector with CAS numbers}
}
\value{
character vector with CAS numbers
}
\description{
Function to convert integer representation of CAS numbers to character representation
}
\author{
Marvin Steijaert
}
