% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substanceUtils.R
\name{intToEC}
\alias{intToEC}
\title{Function to convert integer representation of EC numbers to character representation}
\usage{
intToEC(vec)
}
\arguments{
\item{vec}{integer vector with EC numbers}
}
\value{
character vector with EC numbers
}
\description{
Function to convert integer representation of EC numbers to character representation
}
\author{
Marvin Steijaert
}
