% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPathwayRanking.R
\name{maxWithNA}
\alias{maxWithNA}
\title{Returns NA if only NA values are provided. Otherwise, the maximum of non-NA values is returned}
\usage{
maxWithNA(vec)
}
\arguments{
\item{vec}{numeric vector}
}
\value{
maximum value
}
\description{
Returns NA if only NA values are provided. Otherwise, the maximum of non-NA values is returned
}
