% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPathwayRanking.R
\name{minWithNA}
\alias{minWithNA}
\title{Returns NA if only NA values are provided. Otherwise, the minimum of non-NA values is returned}
\usage{
minWithNA(vec)
}
\arguments{
\item{vec}{numeric vector}
}
\value{
maximum value
}
\description{
Returns NA if only NA values are provided. Otherwise, the minimum of non-NA values is returned
}
