% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modalDialogWithCross.R
\name{modalDialogWithCross}
\alias{modalDialogWithCross}
\title{Modal dialog with cross to close the window
Extension of the shiny::modalDialog()}
\usage{
modalDialogWithCross(
  ...,
  addCross = TRUE,
  title = NULL,
  footer = modalButton("Dismiss"),
  size = c("m", "s", "l"),
  easyClose = FALSE,
  fade = TRUE
)
}
\arguments{
\item{...}{ui main content of the dialog}

\item{addCross}{boolean, whether to add cross button at right top to close the window
the id of button is 'modal-cross'}

\item{title}{character, title for the dialog}

\item{footer}{ui, content for footer}

\item{size}{character, should be one of \code{c("m", "s", "l")}}

\item{easyClose}{boolean, whether the dialog can be closed by escape button}

\item{fade}{boolean, when closing the window using fade effect}
}
\value{
modal dialog is being displayed
}
\description{
Modal dialog with cross to close the window
Extension of the shiny::modalDialog()
}
\author{
mvarewyck
}
