% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPathwayRanking.R
\name{PathwayRanking}
\alias{PathwayRanking}
\title{Class for pathway ranking of a selected substance}
\description{
Class for pathway ranking of a selected substance

Class for pathway ranking of a selected substance
}
\examples{

## ------------------------------------------------
## Method `PathwayRanking$phenotypePathwayByOntology`
## ------------------------------------------------

\dontrun{
ranking <- phenotypePathwayByOntology(
queryPhenotype = c("ZP:0004424","ZP:0008523","ZP:0020346","ZP:0000038","ZP:0001036","ZP:0001211","ZP:0004424"),
go_id_prefix = "ZP",
filterICw = NULL
)
}

## ------------------------------------------------
## Method `PathwayRanking$createCasePhenotype`
## ------------------------------------------------

\dontrun{
self$createCasePhenotype(queryPhenotype = c("ZP:0004424","ZP:0008523"), "Danio rerio", 1)
}

## ------------------------------------------------
## Method `PathwayRanking$matchProfile`
## ------------------------------------------------

\dontrun{
self$matchProfile(profile, querySpecies)
}

## ------------------------------------------------
## Method `PathwayRanking$rankAUC`
## ------------------------------------------------

\dontrun{
self$rankAUC(targetPathway, profile)
}

## ------------------------------------------------
## Method `PathwayRanking$runSimulations`
## ------------------------------------------------

\dontrun{
self$runSimulations(ontology = "ZP", nRuns = 1000, querySizes = c(10,20,30,40))
}

## ------------------------------------------------
## Method `PathwayRanking$estimatePvalue`
## ------------------------------------------------

\dontrun{
self$estimatePvalue(ontology = "ZP", querySize= 10, pathway = "R-HSA-8856828", rankPercent = 12.3)
}

## ------------------------------------------------
## Method `PathwayRanking$estimatePvalueVector`
## ------------------------------------------------

\dontrun{
self$estimatePvalueVector(ontology = "ZP", querySize= 10, pathway = c("R-HSA-8856828","R-HSA-2219528"), rankPercent = c(9.5,12.3))
}

## ------------------------------------------------
## Method `PathwayRanking$plotPvalue`
## ------------------------------------------------

\dontrun{
self$plotPvalue(ontology = "ZP", querySize= 10, pathway = "R-HSA-8856828", rankPercent = 12.3)
}

## ------------------------------------------------
## Method `PathwayRanking$plotSimulationsHistogram`
## ------------------------------------------------

\dontrun{
self$plotSimulationsHistogram(ontology = "ZP", querySize= 10, pathway = "R-HSA-8856828", rankPercent = 12.3)
}
}
\author{
Monique van der Voet

Monique van der Voet

Monique van der Voet

Monique van der Voet

Monique van der Voet, Marvin Steijaert

Marvin Steijaert

Marvin Steijaert

Marvin Steijaert

Marvin Steijaert

Marvin Steijaert
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{classname}}{Name of the class}

\item{\code{database}}{Reference to DartDB object}

\item{\code{substanceid}}{Identifier of substance(s) in self$database}

\item{\code{mammalianPhenotypes}}{Field with all observed mammalian phenotypes for the substance(s)}

\item{\code{nonMammalianPhenotypes}}{Field with all observed non-mammalian phenotypes for the substance(s)}

\item{\code{allInVitro}}{Field with all in vitro data for the substance(s)}

\item{\code{humanGenesList}}{Human pathway data}

\item{\code{pathwayLevels}}{Level(s) of selected human pathways}

\item{\code{pathwaySummaryLevels}}{Level(s) of selected human pathways used for summary}

\item{\code{sizePathwayMin}}{Minimum number of genes in selected human pathways}

\item{\code{sizePathwayMax}}{Maximum number of genes in selected human pathways}

\item{\code{cache}}{Cache for pathway ranking with different data selections}

\item{\code{defaultSelection}}{default data selections}

\item{\code{mammalianPhenotypeOntologies}}{Supported ontologies for mammalian phenotypes}

\item{\code{nonMammalianPhenotypeOntologies}}{Supported ontologies for non-mammalian phenotypes}

\item{\code{phenotypeRankingLowestLevel}}{Logical, indicates if phenotype-based ranking should use lowest level + aggregation}

\item{\code{simulationResults}}{Simulation results used for p-value estimation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PathwayRanking$new()}}
\item \href{#method-.rankPathways}{\code{PathwayRanking$.rankPathways()}}
\item \href{#method-rankPathways}{\code{PathwayRanking$rankPathways()}}
\item \href{#method-phenotypePathwayByOntology}{\code{PathwayRanking$phenotypePathwayByOntology()}}
\item \href{#method-createCasePhenotype}{\code{PathwayRanking$createCasePhenotype()}}
\item \href{#method-createProfile}{\code{PathwayRanking$createProfile()}}
\item \href{#method-matchProfile}{\code{PathwayRanking$matchProfile()}}
\item \href{#method-rankAUC}{\code{PathwayRanking$rankAUC()}}
\item \href{#method-runSimulations}{\code{PathwayRanking$runSimulations()}}
\item \href{#method-estimatePvalue}{\code{PathwayRanking$estimatePvalue()}}
\item \href{#method-estimatePvalueVector}{\code{PathwayRanking$estimatePvalueVector()}}
\item \href{#method-plotPvalue}{\code{PathwayRanking$plotPvalue()}}
\item \href{#method-plotSimulationsHistogram}{\code{PathwayRanking$plotSimulationsHistogram()}}
\item \href{#method-clone}{\code{PathwayRanking$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new PathwayRanking object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayRanking$new(
  database,
  substanceid,
  pathwayLevels = NULL,
  pathwaySummaryLevels = if (length(pathwayLevels)) min(pathwayLevels),
  sizePathwayMin = 0,
  sizePathwayMax = Inf,
  phenotypeRankingLowestLevel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{database}}{DartDB object}

\item{\code{substanceid}}{selected substanceid in database}

\item{\code{pathwayLevels}}{optional level(s) of selected pathways}

\item{\code{pathwaySummaryLevels}}{Level(s) of selected human pathways used for summary}

\item{\code{sizePathwayMin}}{minimum number of genes in selected HUMAN pathways}

\item{\code{sizePathwayMax}}{maximum number of genes in selected HUMAN pathways}

\item{\code{phenotypeRankingLowestLevel}}{logical, indicates if phenotype-based ranking should use lowest level + aggregation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.rankPathways"></a>}}
\if{latex}{\out{\hypertarget{method-.rankPathways}{}}}
\subsection{Method \code{.rankPathways()}}{
Calculates pathway ranking (without using cache)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayRanking$.rankPathways(
  inVitroSelection,
  mammalianPhenoSelection,
  nonMammalianPhenoSelection
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inVitroSelection}}{logical vector indicating which rows in self$allInVitro should be used}

\item{\code{mammalianPhenoSelection}}{logical vector indicating which rows in self$mammalianPhenotypes should be used}

\item{\code{nonMammalianPhenoSelection}}{logical vector indicating which rows in self$nonMammalianPhenotypes should be used}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rankPathways"></a>}}
\if{latex}{\out{\hypertarget{method-rankPathways}{}}}
\subsection{Method \code{rankPathways()}}{
Calculates pathway ranking (using cache for speedup)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayRanking$rankPathways(
  inVitroSelection = self$defaultSelection$inVitro,
  mammalianPhenoSelection = self$defaultSelection$mammalianPhenotypes,
  nonMammalianPhenoSelection = self$defaultSelection$nonMammalianPhenotypes
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inVitroSelection}}{(optional) logical vector indicating which rows in self$allInVitro should be used}

\item{\code{mammalianPhenoSelection}}{logical vector indicating which rows in self$mammalianPhenotypes should be used}

\item{\code{nonMammalianPhenoSelection}}{logical vector indicating which rows in self$nonMammalianPhenotypes should be used}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-phenotypePathwayByOntology"></a>}}
\if{latex}{\out{\hypertarget{method-phenotypePathwayByOntology}{}}}
\subsection{Method \code{phenotypePathwayByOntology()}}{
variant of phenotypePathway that allows specifying ontology instead of species
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayRanking$phenotypePathwayByOntology(
  queryPhenotype,
  go_id_prefix,
  filterICw = NULL,
  lowestLevelOnly = FALSE,
  pValues = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{queryPhenotype}}{list of phenotypes, e.g. c("ZP:0004424","ZP:0008523")}

\item{\code{go_id_prefix}}{prefix of ontology as used in queryPhenotype. (e.g., "ZP", "MP")}

\item{\code{filterICw}}{number that scales the meanICweight, e.g. 1}

\item{\code{lowestLevelOnly}}{Only include lowest level genes for pathways}

\item{\code{pValues}}{If TRUE, p-value estimaes are calculated.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list with data.tables for ranking and contributing phenotypes per pathway
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
ranking <- phenotypePathwayByOntology(
queryPhenotype = c("ZP:0004424","ZP:0008523","ZP:0020346","ZP:0000038","ZP:0001036","ZP:0001211","ZP:0004424"),
go_id_prefix = "ZP",
filterICw = NULL
)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createCasePhenotype"></a>}}
\if{latex}{\out{\hypertarget{method-createCasePhenotype}{}}}
\subsection{Method \code{createCasePhenotype()}}{
Create a test case from a list of phenotypes (phenotype>gene>pathway)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayRanking$createCasePhenotype(
  queryPhenotype,
  querySpecies = NULL,
  filterICw = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{queryPhenotype}}{list of phenotypes, e.g. c("ZP:0004424","ZP:0008523")}

\item{\code{querySpecies}}{(optional) character string with species, e.g. "Caenorhabditis elegans"}

\item{\code{filterICw}}{Filter to use for IC, should be, NULL (no filtering) "mean" or number 0-1.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
casePheno data.table with GO_ID and ICweight
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
self$createCasePhenotype(queryPhenotype = c("ZP:0004424","ZP:0008523"), "Danio rerio", 1)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createProfile"></a>}}
\if{latex}{\out{\hypertarget{method-createProfile}{}}}
\subsection{Method \code{createProfile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayRanking$createProfile(
  casePheno,
  querySpecies = NULL,
  go_id_prefix = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{casePheno}}{data.table with go_id and ICweight}

\item{\code{querySpecies}}{Full scientic name of species (e.g. "Homo Sapiens")}

\item{\code{go_id_prefix}}{Prefix of go_id, used for filtering instead of querySpecies.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a signature profile of phenotypes, a list of genes with at least one match, counting the total matches
}

\subsection{Returns}{
profile data.table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-matchProfile"></a>}}
\if{latex}{\out{\hypertarget{method-matchProfile}{}}}
\subsection{Method \code{matchProfile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayRanking$matchProfile(profile, querySpecies, lowestLevelOnly = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{profile}}{data.table from class method createProfile}

\item{\code{querySpecies}}{character string with species, e.g. "Caenorhabditis elegans"}

\item{\code{lowestLevelOnly}}{Only include lowest level genes for pathways}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Retrieve list of pathways that include at least one of the genes
}

\subsection{Returns}{
targetPathway
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
self$matchProfile(profile, querySpecies)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rankAUC"></a>}}
\if{latex}{\out{\hypertarget{method-rankAUC}{}}}
\subsection{Method \code{rankAUC()}}{
Rank pathways using AUC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayRanking$rankAUC(targetPathway, profile)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{targetPathway}}{data.table from class method matchProfile}

\item{\code{profile}}{data.table from class method createProfile}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ranking
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
self$rankAUC(targetPathway, profile)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runSimulations"></a>}}
\if{latex}{\out{\hypertarget{method-runSimulations}{}}}
\subsection{Method \code{runSimulations()}}{
Run Monte Carlo simulations for phenotypePathwayByOntology
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayRanking$runSimulations(
  ontology,
  nRuns = 1000,
  querySizes = c(10, 20, 30, 40),
  mc.cores = max(parallel::detectCores() - 1, 1),
  digits = 2L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ontology}}{name (prefix of phenotype identifiers) of ontology}

\item{\code{nRuns}}{number of simulations}

\item{\code{querySizes}}{Vector with sizes of phenotype queries}

\item{\code{mc.cores}}{Passed to parallel::mcmapply.}

\item{\code{digits}}{Integer value passed to round(). If NULL or non-numeric, rounding will be skipped.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Use rounding (digits parameter) to reduce the size on disk and time to load.
}

\subsection{Returns}{
data.table with simulation results
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
self$runSimulations(ontology = "ZP", nRuns = 1000, querySizes = c(10,20,30,40))
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimatePvalue"></a>}}
\if{latex}{\out{\hypertarget{method-estimatePvalue}{}}}
\subsection{Method \code{estimatePvalue()}}{
Estimate p-value using Monte Carlo simulations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayRanking$estimatePvalue(ontology, querySize, pathway, rankPercent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ontology}}{Name (prefix of phenotype identifiers) of ontology}

\item{\code{querySize}}{Number of phenotypes in query}

\item{\code{pathway}}{Pathway identifier}

\item{\code{rankPercent}}{Original rankPercent score returned by self$rankAUC}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
p-value
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
self$estimatePvalue(ontology = "ZP", querySize= 10, pathway = "R-HSA-8856828", rankPercent = 12.3)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimatePvalueVector"></a>}}
\if{latex}{\out{\hypertarget{method-estimatePvalueVector}{}}}
\subsection{Method \code{estimatePvalueVector()}}{
Estimate vector of p-values using Monte Carlo simulations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayRanking$estimatePvalueVector(
  ontology,
  querySize,
  pathway,
  rankPercent,
  skipNoSimulations = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ontology}}{Name (prefix of phenotype identifiers) of ontology}

\item{\code{querySize}}{Number of phenotypes in query}

\item{\code{pathway}}{Vector with pathway identifiers}

\item{\code{rankPercent}}{Vector with original rankPercent scores returned by self$rankAUC}

\item{\code{skipNoSimulations}}{Boolean, if TRUE, pathways without simulation results will get an NA value. If FALSE, a pvalue will be calculated (always lowest possible or 1 for rankPercent==0).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector with p-values
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
self$estimatePvalueVector(ontology = "ZP", querySize= 10, pathway = c("R-HSA-8856828","R-HSA-2219528"), rankPercent = c(9.5,12.3))
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPvalue"></a>}}
\if{latex}{\out{\hypertarget{method-plotPvalue}{}}}
\subsection{Method \code{plotPvalue()}}{
Plot simulation-based p-value estimate for different number of simulations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayRanking$plotPvalue(ontology, querySize, pathway, rankPercent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ontology}}{Name (prefix of phenotype identifiers) of ontology}

\item{\code{querySize}}{Number of phenotypes in query}

\item{\code{pathway}}{Pathway identifier}

\item{\code{rankPercent}}{Original rankPercent score returned by self$rankAUC}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
self$plotPvalue(ontology = "ZP", querySize= 10, pathway = "R-HSA-8856828", rankPercent = 12.3)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSimulationsHistogram"></a>}}
\if{latex}{\out{\hypertarget{method-plotSimulationsHistogram}{}}}
\subsection{Method \code{plotSimulationsHistogram()}}{
Plot histogram with simulated rankPercent scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayRanking$plotSimulationsHistogram(
  ontology,
  querySize,
  pathway,
  rankPercent = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ontology}}{Name (prefix of phenotype identifiers) of ontology}

\item{\code{querySize}}{Number of phenotypes in query}

\item{\code{pathway}}{Pathway identifier}

\item{\code{rankPercent}}{Original rankPercent score returned by self$rankAUC (optional). Will be shown with a red line.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
self$plotSimulationsHistogram(ontology = "ZP", querySize= 10, pathway = "R-HSA-8856828", rankPercent = 12.3)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PathwayRanking$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
