% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPhenoPathway.R
\name{PhenoPathway}
\alias{PhenoPathway}
\title{Class for dealing with Phenotype and Pathway data}
\description{
Class for dealing with Phenotype and Pathway data

Class for dealing with Phenotype and Pathway data
}
\details{
For reasons of code visibility, the different components are implemented in a series of parent classes
}
\examples{

## ------------------------------------------------
## Method `PhenoPathway$getPathwayOrthology`
## ------------------------------------------------

\dontrun{
self$getPathwayOrthology(pathwayid = "R-HSA-2029481")
}
}
\author{
Marvin Steijaert

Marvin Steijaert

Marvin Steijaert

Marvin Steijaert

Marvin Steijaert

Monique van der Voet

Marvin Steijaert

Marvin Steijaert
}
\section{Super classes}{
\code{\link[dartTools:Data]{dartTools::Data}} -> \code{\link[dartTools:Substances]{dartTools::Substances}} -> \code{PhenoPathway}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{classname}}{Name of the class}

\item{\code{humanPathwayLevelsCache}}{Cache data.table object with pathway level mapping}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-getHumanPathwayLevels}{\code{PhenoPathway$getHumanPathwayLevels()}}
\item \href{#method-getHumanPathwayParents}{\code{PhenoPathway$getHumanPathwayParents()}}
\item \href{#method-getHumanPathwaysForLevel}{\code{PhenoPathway$getHumanPathwaysForLevel()}}
\item \href{#method-getHumanPathways}{\code{PhenoPathway$getHumanPathways()}}
\item \href{#method-getHumanPathwayGenes}{\code{PhenoPathway$getHumanPathwayGenes()}}
\item \href{#method-retrieveTaxon}{\code{PhenoPathway$retrieveTaxon()}}
\item \href{#method-getPathwayOrthology}{\code{PhenoPathway$getPathwayOrthology()}}
\item \href{#method-mapGenesToHumanGenes}{\code{PhenoPathway$mapGenesToHumanGenes()}}
\item \href{#method-clone}{\code{PhenoPathway$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="addData">}\href{../../dartTools/html/Data.html#method-addData}{\code{dartTools::Data$addData()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="addTable">}\href{../../dartTools/html/Data.html#method-addTable}{\code{dartTools::Data$addTable()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="checkTableConventions">}\href{../../dartTools/html/Data.html#method-checkTableConventions}{\code{dartTools::Data$checkTableConventions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="dump">}\href{../../dartTools/html/Data.html#method-dump}{\code{dartTools::Data$dump()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="getData">}\href{../../dartTools/html/Data.html#method-getData}{\code{dartTools::Data$getData()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="initialize">}\href{../../dartTools/html/Data.html#method-initialize}{\code{dartTools::Data$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="loadDump">}\href{../../dartTools/html/Data.html#method-loadDump}{\code{dartTools::Data$loadDump()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="ncols">}\href{../../dartTools/html/Data.html#method-ncols}{\code{dartTools::Data$ncols()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="nrows">}\href{../../dartTools/html/Data.html#method-nrows}{\code{dartTools::Data$nrows()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="print">}\href{../../dartTools/html/Data.html#method-print}{\code{dartTools::Data$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="printAllTables">}\href{../../dartTools/html/Data.html#method-printAllTables}{\code{dartTools::Data$printAllTables()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Data" data-id="writeTxtTable">}\href{../../dartTools/html/Data.html#method-writeTxtTable}{\code{dartTools::Data$writeTxtTable()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id=".mergeSubstancesOnce">}\href{../../dartTools/html/Substances.html#method-.mergeSubstancesOnce}{\code{dartTools::Substances$.mergeSubstancesOnce()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id=".updateSubstanceids">}\href{../../dartTools/html/Substances.html#method-.updateSubstanceids}{\code{dartTools::Substances$.updateSubstanceids()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="addConstituents">}\href{../../dartTools/html/Substances.html#method-addConstituents}{\code{dartTools::Substances$addConstituents()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="addIdTypes">}\href{../../dartTools/html/Substances.html#method-addIdTypes}{\code{dartTools::Substances$addIdTypes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="addSubstanceActivity">}\href{../../dartTools/html/Substances.html#method-addSubstanceActivity}{\code{dartTools::Substances$addSubstanceActivity()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="addSubstanceCategory">}\href{../../dartTools/html/Substances.html#method-addSubstanceCategory}{\code{dartTools::Substances$addSubstanceCategory()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="addSubstances">}\href{../../dartTools/html/Substances.html#method-addSubstances}{\code{dartTools::Substances$addSubstances()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getDescriptors">}\href{../../dartTools/html/Substances.html#method-getDescriptors}{\code{dartTools::Substances$getDescriptors()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getMonoSubstanceSmiles">}\href{../../dartTools/html/Substances.html#method-getMonoSubstanceSmiles}{\code{dartTools::Substances$getMonoSubstanceSmiles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getSameCategorySubstances">}\href{../../dartTools/html/Substances.html#method-getSameCategorySubstances}{\code{dartTools::Substances$getSameCategorySubstances()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getSimilarConstituents">}\href{../../dartTools/html/Substances.html#method-getSimilarConstituents}{\code{dartTools::Substances$getSimilarConstituents()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getSimilarSubstances">}\href{../../dartTools/html/Substances.html#method-getSimilarSubstances}{\code{dartTools::Substances$getSimilarSubstances()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getSubstanceActivity">}\href{../../dartTools/html/Substances.html#method-getSubstanceActivity}{\code{dartTools::Substances$getSubstanceActivity()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getSubstanceExtIds">}\href{../../dartTools/html/Substances.html#method-getSubstanceExtIds}{\code{dartTools::Substances$getSubstanceExtIds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="getSubstanceIds">}\href{../../dartTools/html/Substances.html#method-getSubstanceIds}{\code{dartTools::Substances$getSubstanceIds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="inspect">}\href{../../dartTools/html/Substances.html#method-inspect}{\code{dartTools::Substances$inspect()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="mapSmilesToSubstance">}\href{../../dartTools/html/Substances.html#method-mapSmilesToSubstance}{\code{dartTools::Substances$mapSmilesToSubstance()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="maxSubstanceid">}\href{../../dartTools/html/Substances.html#method-maxSubstanceid}{\code{dartTools::Substances$maxSubstanceid()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="mergeSubstances">}\href{../../dartTools/html/Substances.html#method-mergeSubstances}{\code{dartTools::Substances$mergeSubstances()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="tabulateSubstanceCategory">}\href{../../dartTools/html/Substances.html#method-tabulateSubstanceCategory}{\code{dartTools::Substances$tabulateSubstanceCategory()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="tabulateSubstanceInfo">}\href{../../dartTools/html/Substances.html#method-tabulateSubstanceInfo}{\code{dartTools::Substances$tabulateSubstanceInfo()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dartTools" data-topic="Substances" data-id="updateDescriptors">}\href{../../dartTools/html/Substances.html#method-updateDescriptors}{\code{dartTools::Substances$updateDescriptors()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHumanPathwayLevels"></a>}}
\if{latex}{\out{\hypertarget{method-getHumanPathwayLevels}{}}}
\subsection{Method \code{getHumanPathwayLevels()}}{
Get all human pathways, their level and parents
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenoPathway$getHumanPathwayLevels(returnParents = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{returnParents}}{Logical value indicating if parent pathways should be included}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHumanPathwayParents"></a>}}
\if{latex}{\out{\hypertarget{method-getHumanPathwayParents}{}}}
\subsection{Method \code{getHumanPathwayParents()}}{
Get all human pathways upstream of provided pathway identifiers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenoPathway$getHumanPathwayParents(pathwayids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pathwayids}}{Character vector with one or more Reactome pathway identifiers}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector with pathway identifiers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHumanPathwaysForLevel"></a>}}
\if{latex}{\out{\hypertarget{method-getHumanPathwaysForLevel}{}}}
\subsection{Method \code{getHumanPathwaysForLevel()}}{
Get all human pathways at and below a specified level and the corresponding parent pathways at the specified level
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenoPathway$getHumanPathwaysForLevel(level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{Integer specifying a level}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHumanPathways"></a>}}
\if{latex}{\out{\hypertarget{method-getHumanPathways}{}}}
\subsection{Method \code{getHumanPathways()}}{
Function to get all human pathways and apply some filtering
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenoPathway$getHumanPathways(
  levels = NULL,
  sizePathwayMin = 0,
  sizePathwayMax = Inf,
  asVector = FALSE,
  renameEvent = FALSE,
  lowestLevelOnly = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{levels}}{(optional) vector with selected level(s) of pathway in reactome hierarchy}

\item{\code{sizePathwayMin}}{minimum number of genes in pathway, e.g. 4}

\item{\code{sizePathwayMax}}{maximum number of genes in pathway, e.g. 100}

\item{\code{asVector}}{if TRUE, a named vector is returned. If FALSE, a data.table is returned}

\item{\code{renameEvent}}{if TRUE, replicated pathway names will be rename to avoid ambiguity}

\item{\code{lowestLevelOnly}}{logical, indicates if only the genes annotated at the lowest level are included}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
table or vector with Reactome pathways
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHumanPathwayGenes"></a>}}
\if{latex}{\out{\hypertarget{method-getHumanPathwayGenes}{}}}
\subsection{Method \code{getHumanPathwayGenes()}}{
Function to get all human pathways genes and apply some filtering
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenoPathway$getHumanPathwayGenes(
  levels = NULL,
  sizePathwayMin = 0,
  sizePathwayMax = Inf,
  renameEvent = FALSE,
  lowestLevelOnly = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{levels}}{(optional) vector with selected level(s) of pathway in reactome hierarchy}

\item{\code{sizePathwayMin}}{minimum number of genes in pathway, e.g. 4}

\item{\code{sizePathwayMax}}{maximum number of genes in pathway, e.g. 100}

\item{\code{renameEvent}}{if TRUE, replicated pathway names will be rename to avoid ambiguity}

\item{\code{lowestLevelOnly}}{logical, indicates if only the genes annotated at the lowest level are included}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list with data.tables "pathways" and "pathwayGenes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-retrieveTaxon"></a>}}
\if{latex}{\out{\hypertarget{method-retrieveTaxon}{}}}
\subsection{Method \code{retrieveTaxon()}}{
Retrieve taxon id from species name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenoPathway$retrieveTaxon(querySpecies)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{querySpecies}}{character string with species, e.g. "Caenorhabditis elegans"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
queryTaxon
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPathwayOrthology"></a>}}
\if{latex}{\out{\hypertarget{method-getPathwayOrthology}{}}}
\subsection{Method \code{getPathwayOrthology()}}{
get pathway orthology table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenoPathway$getPathwayOrthology(
  pathwayid = NULL,
  description = NULL,
  renameEvent = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pathwayid}}{Reactome pathway identifier}

\item{\code{description}}{Reactome pathway description}

\item{\code{renameEvent}}{passed to self$getHumanPathwayGenes. If TRUE, replicated pathway names will be rename to avoid ambiguity}

\item{\code{...}}{additional parameters passed to self$getHumanPathwayGenes}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Preferably used pathwayid instead of description. Description can be ambiguous
(in rare cases, the same description is used for multiple pathwayids)
}

\subsection{Returns}{
table
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
self$getPathwayOrthology(pathwayid = "R-HSA-2029481")
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapGenesToHumanGenes"></a>}}
\if{latex}{\out{\hypertarget{method-mapGenesToHumanGenes}{}}}
\subsection{Method \code{mapGenesToHumanGenes()}}{
map genes in a table to the human orthologs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenoPathway$mapGenesToHumanGenes(inputTable)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputTable}}{data.table with at least columns 'gene' and 'species'}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Only mouse and rat are currently mapped.
}

\subsection{Returns}{
data.table object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenoPathway$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
