% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appGene.R
\name{plotGeneData}
\alias{plotGeneData}
\title{Bar plot for genes}
\usage{
plotGeneData(
  tableGenes,
  colors = list(dartBlue = "#488df4", remainder = "#488df4", one2one = "#bee126",
    one2many = "#ffde17", many2many = "#488df4"),
  barWidth = NULL,
  width = 600,
  height = NULL
)
}
\arguments{
\item{tableGenes}{table, as returned by \code{\link{summarizeGeneData}}}

\item{colors}{list, colors for the gene classes 
\code{c("one2one", "one2many", "many2many")} (optional)}

\item{barWidth}{bar width (optional)}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
plotly, horizontal stacked barplot
}
\description{
Bar plot for genes
}
\author{
mvarewyck
}
