% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectRdkit.R
\name{processSmilesWithReticulate}
\alias{processSmilesWithReticulate}
\title{Auxiliary function to processes smiles in parallel}
\usage{
processSmilesWithReticulate(
  smiles,
  python_function_name = c("standardize_smiles", "draw_smiles",
    "morgan_fp_bitvector_int", "inchi_to_smiles"),
  timeout = 10,
  nWorkers = parallel::detectCores() - 1,
  strategy = getOption("dartpaths_python_strategy")
)
}
\arguments{
\item{smiles}{Vector with smiles strings.}

\item{python_function_name}{One of the available function names for parallel procesing.}

\item{timeout}{Number of seconds after which a worker is interrupted}

\item{nWorkers}{Number of workers (optional)}

\item{strategy}{Strategy for processing multiple smiles (can be "multiprocessing", "threading" or "sequential")}
}
\value{
processed output of smiles
}
\description{
Auxiliary function to processes smiles in parallel
}
\author{
Marvin Steijaert
}
