% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classData.R
\name{readTxtTable}
\alias{readTxtTable}
\title{Function to read dumped tables from text files}
\usage{
readTxtTable(dumpPath, tablename)
}
\arguments{
\item{dumpPath}{Path of directory containing text file}

\item{tablename}{Name of table (i.e., basename of the text file)}
}
\value{
data.table object
}
\description{
Function to read dumped tables from text files
}
\author{
Marvin Steijaert
}
