% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etlUtils.R
\name{readWebOrFtp}
\alias{readWebOrFtp}
\title{utility function to download files from web and ftp locations and read them as data.table}
\usage{
readWebOrFtp(url, directory, filename = NULL, fileInArchive = NULL, ...)
}
\arguments{
\item{url}{url}

\item{directory}{output directory}

\item{filename}{name out output file (if NULL, this is automatically generated from url)}

\item{fileInArchive}{name of file in zip archive}

\item{...}{Extra parameters passed to the file reading function (data.table::fread() or openxlsx::read.xlsx())}
}
\description{
utility function to download files from web and ftp locations and read them as data.table
}
\author{
Marvin Steijaert
}
