% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substanceUtils.R
\name{removeNonInformativeEntries}
\alias{removeNonInformativeEntries}
\title{Function to remove non-informative rows from a data.table}
\usage{
removeNonInformativeEntries(inputTable, infoColumns)
}
\arguments{
\item{inputTable}{data.table}

\item{infoColumns}{names of columns in in data.table}
}
\value{
list with items inputTable (adjusted version of inputTable) and informative (logical vector indicating if row in original table is informative)
}
\description{
Function to remove non-informative rows from a data.table
}
\details{
Rows in which all specified columns have NA value are removed
}
\author{
Marvin Steijaert
}
