% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbase.R
\name{retrieveDetails}
\alias{retrieveDetails}
\title{Function to help the processing of the retrieved Wormbase data}
\usage{
retrieveDetails(text)
}
\arguments{
\item{text}{The 'details' section of the 'affected genes' table of the molecule.}
}
\value{
The 'details' section of the 'affected genes' table of the molecule in a character vector.
}
\description{
Function to help the processing of the retrieved Wormbase data
}
\author{
Marijke Van Moerbeke
}
