% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDart.R
\name{runDart}
\alias{runDart}
\title{Run the UI for DART}
\usage{
runDart(port = 3838, ...)
}
\arguments{
\item{port}{integer, port to run app on}

\item{...}{further arguments that can be passed to \code{\link[shiny]{runApp}}}
}
\value{
no return value
}
\description{
Run the UI for DART
}
