% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectRdkit.R
\name{smilesToPng}
\alias{smilesToPng}
\title{smilesToPng}
\usage{
smilesToPng(smiles, pythonPath = getOption("dartpaths_python"))
}
\arguments{
\item{smiles}{single SMILES string or a vector with multiple SMILES}

\item{pythonPath}{full path of python executable}
}
\value{
vector with character strings
}
\description{
use rdkit to create png image (html img element with base64 encoded image)
}
\examples{
\dontrun{
smiles <- c("N[C@H](C)C(=O)O","O1CCCC[C@H]1C")
result <- smilesToPng(smiles)
}
}
\author{
Marvin Steijaert
}
