% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDartDB.R
\name{splitToNumericVectors}
\alias{splitToNumericVectors}
\title{Auxialiary function to split character strings to numeric vectors}
\usage{
splitToNumericVectors(vec, split = "|")
}
\arguments{
\item{vec}{Vector with strings}

\item{split}{Character string used for splitting (regular expressions are not supported).}
}
\value{
list with character values
}
\description{
Auxialiary function to split character strings to numeric vectors
}
\details{
can be used as a workaround until data.table::fread supports sep2
see also: https://github.com/Rdatatable/data.table/issues/1162
}
\author{
Marvin Steijaert
}
