% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectRdkit.R
\name{standardizeSmilesRdkit}
\alias{standardizeSmilesRdkit}
\title{standardize smiles}
\usage{
standardizeSmilesRdkit(
  smiles,
  pythonPath = getOption("dartpaths_python"),
  cache = TRUE,
  chunkSize = 1000
)
}
\arguments{
\item{smiles}{single SMILES string or a vector with multiple SMILES}

\item{pythonPath}{full path of python executable}

\item{cache}{If TRUE, a cache file will be used to avoid repeated standardization of the same SMILES}

\item{chunkSize}{Number of smiles per chunk sent to standardizer}
}
\value{
data.table object with standardized smiles
}
\description{
apply Rdkit normalization via reticulate binding to SMILES representation of chemical structures
}
\details{
Rdkit (python) is called to take care of the standardization
}
\examples{
\dontrun{
smiles <- c("N[C@H](C)C(=O)O","O1CCCC[C@H]1C")
result <- standardizeSmiles(smiles)
}
}
\author{
Marvin Steijaert
}
