% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substanceUtils.R
\name{standardizeSpecies}
\alias{standardizeSpecies}
\title{Function to standardize species names}
\usage{
standardizeSpecies(species)
}
\arguments{
\item{species}{Vector with species names}
}
\value{
Vector with standardized species names
}
\description{
Function to standardize species names
}
