% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appGene.R
\name{summarizeGeneData}
\alias{summarizeGeneData}
\title{Create summary table of gene data}
\usage{
summarizeGeneData(geneData)
}
\arguments{
\item{geneData}{data.table object (as returned by \code{DartDB$getPathwayOrthology()})}
}
\value{
table, number of genes of each class
\code{c("0", "ortholog_one2one", "ortholog_one2many", "ortholog_many2many")}
}
\description{
Create summary table of gene data
}
\author{
mvarewyck
}
