% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appCompound.R
\name{uiTable}
\alias{uiTable}
\title{Create (interactive) table for UI in a unified format}
\usage{
uiTable(
  myData,
  columnMap = DEFAULTCOLUMNMAP,
  columnSelection = names(columnMap),
  columnConversions = COLUMNCONVERSIONS,
  interactive = FALSE,
  escape = FALSE,
  scrollX = TRUE,
  filter = "none",
  container = NULL,
  options = NULL,
  extensions = NULL,
  selection = "single",
  callback = JS("return table;")
)
}
\arguments{
\item{myData}{data.table}

\item{columnMap}{named character vector mapping original column names to displayed column names}

\item{columnSelection}{columns that should be displayed (if available)}

\item{columnConversions}{named vector with functions to convert values in displayed table}

\item{interactive}{indicates wether the datatable should be interactive}

\item{escape}{passed to \code{datatable()}}

\item{scrollX}{passed to \code{datatable()}}

\item{filter}{passed to \code{datatable()}}

\item{container}{passed to \code{datatable()}}

\item{options}{passed to \code{datatable()}}

\item{extensions}{passed to \code{datatable()}}

\item{selection}{passed to \code{datatable()}}

\item{callback}{passed to \code{datatable()}}
}
\value{
datatable, interactive version of \code{myData}
}
\description{
Create (interactive) table for UI in a unified format
}
\details{
Unified format: same column names and same column order, hide non-informative columns
}
\author{
Marvin Steijaert
}
