% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{wormbaseAceElementParser}
\alias{wormbaseAceElementParser}
\title{helper function for wormbaseAceMoleculeParser}
\usage{
wormbaseAceElementParser(element)
}
\arguments{
\item{element}{element from wormbase ace file}
}
\value{
data.table object
}
\description{
helper function for wormbaseAceMoleculeParser
}
\author{
Marvin Steijaert
}
