% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{wormbaseAceMoleculeParser}
\alias{wormbaseAceMoleculeParser}
\title{Parser for Wormbase ace file with Molecule data}
\usage{
wormbaseAceMoleculeParser(filepath)
}
\arguments{
\item{filepath}{Path of .ace file}
}
\value{
data.table object
}
\description{
Parser for Wormbase ace file with Molecule data
}
\details{
Creates a data.table that can be fed to addSubstances method of Substances class
}
\author{
Marvin Steijaert
}
