% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbase.R
\name{wormbaseAPI}
\alias{wormbaseAPI}
\title{Function connecting to the wormbase database for specific molecule}
\usage{
wormbaseAPI(
  prefix = NULL,
  identifier = NULL,
  suffix = NULL,
  baseurl = "http://rest.wormbase.org/rest"
)
}
\arguments{
\item{prefix}{A character string to complete the url. Defaults to NULL.}

\item{identifier}{Wormbase identifier for a molecule. Defaults to NULL.}

\item{suffix}{A character string to complete the url. Defaults to NULL.}

\item{baseurl}{The wormbase base url: "http://rest.wormbase.org/rest".}
}
\value{
A processed JSON object with the required information of the identifier.
}
\description{
Function connecting to the wormbase database for specific molecule
}
\examples{
\dontrun{
result <- wormbaseAPI(prefix = "widget/molecule", identifier ="WBMol:00003509", suffix = "affected")
}
}
\author{
Marijke Van Moerbeke
}
