% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbase.R
\name{wormbaseDataRetrieval}
\alias{wormbaseDataRetrieval}
\title{Function to retrieve and process the 'affected genes' of a vector of molecules in the Wormbase database}
\usage{
wormbaseDataRetrieval(molecules = NULL)
}
\arguments{
\item{molecules}{A character vector of the molecule to retrieve. Defaults to NULL.}
}
\value{
A data.table containing the processed 'affected genes' of the molecules.
}
\description{
Function to retrieve and process the 'affected genes' of a vector of molecules in the Wormbase database
}
\examples{
\donttest{
result <- wormbaseDataRetrieval(molecules = c("WBMol:00000280", "WBMol:00000373", "WBMol:00000896"))
}
}
\author{
Marijke Van Moerbeke
}
