% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbase.R
\name{wormbaseMoleculeData}
\alias{wormbaseMoleculeData}
\title{Function to retrieve and process the 'affected genes' of a single molecule in the Wormbase database}
\usage{
wormbaseMoleculeData(molecule = NULL, verbose = FALSE)
}
\arguments{
\item{molecule}{The molecule to retrieve. Defaults to NULL.}

\item{verbose}{Logical value to switch on/off extra verbosity.}
}
\value{
A data.table containing the processed 'affected genes' table of the molecule.
}
\description{
Function to retrieve and process the 'affected genes' of a single molecule in the Wormbase database
}
\author{
Marijke Van Moerbeke
}
