% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbase.R
\name{wormBaseSmilesRetrieval}
\alias{wormBaseSmilesRetrieval}
\title{Function to retrieve and process the smiles of a vector of molecules in the Wormbase database}
\usage{
wormBaseSmilesRetrieval(molecules)
}
\arguments{
\item{molecules}{A character vector of the molecule to retrieve.}
}
\value{
Vector with the smiles retrieved for the molecules (NA if missing).
}
\description{
Function to retrieve and process the smiles of a vector of molecules in the Wormbase database
}
\examples{
\donttest{
result <- wormBaseSmilesRetrieval(molecules = c("WBMol:00000001", "WBMol:00000002", "WBMol:00000003"))
}
}
\author{
Marvin Steijaert
}
