% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectRdkit.R
\name{inchiToSmiles}
\alias{inchiToSmiles}
\title{inchiToSmiles}
\usage{
inchiToSmiles(inchis, pythonPath = getOption("dartpaths_python"))
}
\arguments{
\item{inchis}{single inchi string or a vector with multiple inchis}

\item{pythonPath}{full path of python executable}
}
\value{
vector with character strings
}
\description{
use rdkit to convert a vector of inchi to a vector of smiles
}
\examples{
\dontrun{
inchis <- c("InChI=1S/C2H6O/c1-2-3/h3H,2H2,1H3","InChI=1S/C6H8O6/c7-1-2(8)5-3(9)4(10)6(11)12-5/h2,5,7-10H,1H2/t2-,5+/m0/s1")
result <- inchiToSmiles(inchis)
}
}
\author{
Marvin Steijaert
}
